Add operator graph optimizer.
The optimizer merges together operators at build time in order to
reduce simulation time for many types of common models.

The optimizer required some modifications to existing graph functions.
In particular, `reverse_edges` was changed to return the complete
graph (the old implementation would remove nodes without edges).
Also, we now consistently use sets for collections of graph edges.

This commit includes some internal tests, as well as an integration
test ensuring that complex models are simulated exactly the same.
If we uncover any models that do change with the optimizer in the
future, it is easy to add tests for those cases.