Rename network inputs.
This makes the product and circular convolution network adhere to the
newly defined standard for network inputs which should be named
'input_*'.

This does not touch the names of inputs to the SPA networks as most of
them will be refactored in nengo_spa anyways.

The old names are still accessible. For technical reasons, they do not
raise a DeprecationWarning, though that would have been preferable.

Addresses #887.