Add a config system quirks notebook
It's confusing that config context, not network membership,
determines the defaults used by the config system. This new notebook
explains that this happens, and gives some insight into why.

This commit also adds a quirks folder to the set of examples, which
could be a good place to put other potentially confusing aspects
of Nengo.

Addresses #615.