Add Samples distribution
This is essentially a no-op distribution that takes a set of
samples and provides them when the `samples` method is called.
It's intended to simplify situations in which a distribution or
samples could be provided by converting samples to a distribution.

The main place this happens is when passing `Ensemble` arguments
to the `EnsembleArray` constructor. Previously, a distribution had
to be provided, since the passed parameters are set as defaults
on the `EnsembleArray` network. Now, any samples passed in are
wrapped in the `Samples` distribution, which makes it possible to
pass in samples. Tests have been added to verify this.

Addresses #691 and #766.