Raise exception on invalid gains.
This also requires fixing the calculation of gains and biases for
sigmoid neurons. Because the intercepts for sigmoid neurons are actually
not an intercept, but the inflection point, valid intercepts depend on
whether the max firing rate is below or above the inflection point.
This commit also adjusts the default refractory time constant of the
sigmoid neurons to play well with intercepts < 1 and max firing rates
between 200 and 400Hz. This required some slight adjustments to tests.

Fixes #1212.