Change DotInc to ElementwiseInc in PES
As defined, the arguments to DotInc should be a matrix and a
vector. In this case it was being called with a scalar
and a vector as input, which would more appropriately
be an ElementwiseInc.

The swap works fine in the reference backend, because
`dot(scalar, vector) == scalar * vector` in NumPy. But in other
backends, such as nengo_ocl or nengo_dl, there may be specialized
implementations for `DotInc` that are optimized under the assumption
that it will be called with a matrix. Since there's no downside to
switching this operation to an ElementwiseInc for the reference
backend, we might as well in order to be consistent.